refactor: require game write for endpoints
