+++ MehrakBot/Services/Dashboard/Mehrak.Dashboard/CodeRedeem/CodesController.cs
@@ -1,4 +1,5 @@
 ﻿using Mehrak.Dashboard.CodeRedeem.Models;
+using Mehrak.Dashboard.Shared;
 using Mehrak.Domain.Shared.Enums;
 using Mehrak.Infrastructure.CodeRedeem;
 using Mehrak.Infrastructure.CodeRedeem.Models;
@@ -11,7 +12,7 @@ namespace Mehrak.Dashboard.CodeRedeem;
 [ApiController]
 [Authorize]
 [Route("codes")]
-public sealed class CodesController : ControllerBase
+public sealed class CodesController : GameWriteController
 {
     private readonly CodeRedeemDbContext m_CodeContext;
     private readonly ILogger<CodesController> m_Logger;
@@ -23,13 +24,14 @@ public sealed class CodesController : ControllerBase
     }
 
     [HttpPatch("add")]
+    [Authorize(Policy = "RequireGameWrite")]
     public async Task<IActionResult> AddCodes([FromQuery] string game, [FromBody] AddCodesRequest request)
     {
         if (!ModelState.IsValid)
             return ValidationProblem(ModelState);
 
         if (!TryParseGame(game, out var parsedGame, out var errorResult))
-            return errorResult!;
+            return BadRequest(new { error = errorResult });
 
         if (!HasGameWriteAccess(game))
             return Forbid();
@@ -47,13 +49,14 @@ public sealed class CodesController : ControllerBase
     }
 
     [HttpDelete("remove")]
+    [Authorize(Policy = "RequireGameWrite")]
     public async Task<IActionResult> RemoveCodes([FromQuery] string game, [FromQuery] RemoveCodesRequest request)
     {
         if (!ModelState.IsValid)
             return ValidationProblem(ModelState);
 
         if (!TryParseGame(game, out var parsedGame, out var errorResult))
-            return errorResult!;
+            return BadRequest(new { error = errorResult });
 
         if (!HasGameWriteAccess(game))
             return Forbid();
@@ -74,38 +77,13 @@ public sealed class CodesController : ControllerBase
     public async Task<IActionResult> ListCodes([FromQuery] string game)
     {
         if (!TryParseGame(game, out var parsedGame, out var errorResult))
-            return errorResult!;
+            return BadRequest(new { error = errorResult });
 
         if (!HasGameWriteAccess(game))
             return Forbid();
 
         var codes = await m_CodeContext.Codes.AsNoTracking().Where(x => x.Game == parsedGame).Select(x => x.Code).ToListAsync();
         return Ok(new { game = parsedGame.ToString(), codes });
-    }
-
-    private static bool TryParseGame(string gameValue, out Game game, out IActionResult? error)
-    {
-        error = null;
-        if (string.IsNullOrWhiteSpace(gameValue) || !Enum.TryParse(gameValue, true, out game))
-        {
-            game = default;
-            error = new BadRequestObjectResult(new { error = "Invalid game value." });
-            return false;
-        }
-
-        return true;
-    }
-
-    private bool HasGameWriteAccess(string game)
-    {
-        if (User.IsInRole("superadmin"))
-            return true;
-
-        var normalized = game.Trim().ToLowerInvariant();
-        var claimValue = $"game_write:{normalized}";
-        return User.Claims.Any(c =>
-            string.Equals(c.Type, "perm", StringComparison.OrdinalIgnoreCase) &&
-            string.Equals(c.Value, claimValue, StringComparison.OrdinalIgnoreCase));
     }
 
     private static List<string> NormalizeCodes(IEnumerable<string> codes)
