+++ MehrakBot/Services/Dashboard/Mehrak.Dashboard/Shared/GameWriteController.cs
@@ -1,0 +1,39 @@
+using Mehrak.Domain.Shared.Enums;
+using Microsoft.AspNetCore.Mvc;
+
+namespace Mehrak.Dashboard.Shared;
+
+[ApiController]
+public abstract class GameWriteController : ControllerBase
+{
+    protected bool HasGameWriteAccess(string game)
+    {
+        if (User.IsInRole("superadmin"))
+            return true;
+
+        var normalized = game.Trim().ToLowerInvariant();
+        var claimValue = $"game_write:{normalized}";
+        return User.Claims.Any(c =>
+            string.Equals(c.Type, "perm", StringComparison.OrdinalIgnoreCase) &&
+            string.Equals(c.Value, claimValue, StringComparison.OrdinalIgnoreCase));
+    }
+
+    protected static bool TryParseGame(string? input, out Game game, out string error)
+    {
+        if (string.IsNullOrWhiteSpace(input))
+        {
+            error = "Game parameter is required.";
+            game = default;
+            return false;
+        }
+
+        if (!Enum.TryParse(input, true, out game))
+        {
+            error = "Invalid game parameter.";
+            return false;
+        }
+
+        error = string.Empty;
+        return true;
+    }
+}
